PluginABI: Create a Functor
We need to be able to do callbacks from various scripting languages
and many/most of them will need to be able to attach extra data to
a callback function, such as the scripting lanugage context and any
variables captured by the callback closure.

This will also be useful later for identifing ownership of function
calls, and allocating memory with the correct allocator.